Handle query parameter, empty search
